Refactor to allow customization of affected and focused spans
