+++ static/app/components/events/interfaces/performance/utils.tsx
@@ -1,7 +1,13 @@
 import * as Sentry from '@sentry/react';
 import keyBy from 'lodash/keyBy';
 
-import {EntrySpans, EntryType, EventTransaction, IssueCategory} from 'sentry/types';
+import {
+  EntrySpans,
+  EntryType,
+  EventTransaction,
+  IssueCategory,
+  IssueType,
+} from 'sentry/types';
 
 import {RawSpanType} from '../spans/types';
 
@@ -47,4 +53,45 @@ export function getSpanInfoFromTransactionEvent(
     offendingSpans: offendingSpanIDs.map(spanID => spansById[spanID]),
     causeSpans: causeSpanIDs.map(spanID => spansById[spanID]),
   };
+}
+
+/**
+ * Given an event for a performance issue, returns the `affectedSpanIds` and `focusedSpanIds`.
+ * Both of these subsets of spans are used to determine which spans are initially visible on the span tree on the issue details
+ * page. The main difference is that the former will be highlighted in red, these spans are intended to indicate the 'root cause' spans
+ * of the issue, with the latter being supplemental spans that are involved in the issue but not necessarily the cause of it.
+ *
+ * @param event
+ */
+export function getProblemSpansForSpanTree(event: EventTransaction): {
+  affectedSpanIds: string[];
+  focusedSpanIds: string[];
+} {
+  const issueType = event?.perfProblem?.issueType;
+  const affectedSpanIds: string[] = [];
+  const focusedSpanIds: string[] = [];
+
+  // By default, offender spans will always be `affected spans`
+  const offenderSpanIds = event.perfProblem?.offenderSpanIds ?? [];
+  affectedSpanIds.push(...offenderSpanIds);
+
+  if (issueType !== IssueType.PERFORMANCE_N_PLUS_ONE_API_CALLS) {
+    const parentSpanIds = event.perfProblem?.parentSpanIds ?? [];
+    affectedSpanIds.push(...parentSpanIds);
+  }
+
+  if (issueType === IssueType.PERFORMANCE_CONSECUTIVE_DB_QUERIES) {
+    const consecutiveSpanIds = event?.perfProblem?.causeSpanIds ?? [];
+
+    if (consecutiveSpanIds.length < 11) {
+      focusedSpanIds.push(...consecutiveSpanIds);
+    }
+  }
+
+  if (issueType === IssueType.PERFORMANCE_N_PLUS_ONE_DB_QUERIES) {
+    const precedingSpans = event.perfProblem?.causeSpanIds ?? [];
+    focusedSpanIds.push(...precedingSpans);
+  }
+
+  return {affectedSpanIds, focusedSpanIds};
 }
